requirements.txt
crispy-bootstrap4
Django
django-crispy-forms
fontawesomefree
numpy
torch
torchvision
tqdm
transformers
Pillow
opencv-python
scikit-image
split-folders
